Upgrade ActiveSupport to 3.2.19+
This loosens the restriction on active support. We were doing exact
because it was blocking security patch upgrades in other projects. The
version and should have been `~>` on active support.

I also upgraded rspec to 3.0 and updated the spec syntax to `expect`
instead of `should`. As well as got rid of deprication warnings.