Removes dependency on `active_support`
This required a couple of different approaches. First, the `.blank?`
checks have been converted to a `.to_s.empty?` check. This will cover
everything other than the case of `false`. Second, the reverse_merge was
switched around to work the same way without bringing in the entire
behemoth that is activesupport.

Lastly, I added a runtime dependency for a specific version of i18n and
also setup the available locales per the update in the spec helper.